feat: clean test as in docs
