+++ src/components/RecipeCard.test.tsx
@@ -1,34 +1,8 @@
-import { ThemeProvider } from "@mui/material";
-import { AppRouterCacheProvider } from "@mui/material-nextjs/v15-appRouter";
-import {
-  render,
-  RenderOptions,
-  RenderResult,
-  screen,
-} from "@testing-library/react";
+import { render, screen } from "@/test-utils";
 
 import "@testing-library/jest-dom";
 
-import { ReactElement } from "react";
-
-import { theme } from "@/lib/MUITheme";
-
 import { RecipeCard } from "./RecipeCard";
-
-const Providers = ({
-  children,
-}: {
-  children: React.ReactNode;
-}): ReactElement => (
-  <AppRouterCacheProvider>
-    <ThemeProvider theme={theme}>{children}</ThemeProvider>
-  </AppRouterCacheProvider>
-);
-
-const customRender = (
-  ui: ReactElement,
-  options?: Omit<RenderOptions, "wrapper">,
-): RenderResult => render(ui, { wrapper: Providers, ...options });
 
 const mockRecipe = {
   id: "1",
@@ -39,7 +13,7 @@ const mockRecipe = {
 
 describe("RecipeCard", () => {
   it("should render the recipe name and image", () => {
-    customRender(<RecipeCard recipe={mockRecipe} imageSizes="100vw" />);
+    render(<RecipeCard recipe={mockRecipe} imageSizes="100vw" />);
 
     expect(screen.getByText("Delicious Dish")).toBeInTheDocument();
 
@@ -52,7 +26,7 @@ describe("RecipeCard", () => {
   });
 
   it("should render an image even if none is provided", () => {
-    customRender(
+    render(
       <RecipeCard
         recipe={{ ...mockRecipe, image_url: undefined }}
         imageSizes="100vw"
@@ -68,7 +42,7 @@ describe("RecipeCard", () => {
   });
 
   it("should render an alternative text even if none is provided", () => {
-    customRender(
+    render(
       <RecipeCard
         recipe={{ ...mockRecipe, name: undefined }}
         imageSizes="100vw"
@@ -85,16 +59,14 @@ describe("RecipeCard", () => {
       description: "A".repeat(200),
     };
 
-    customRender(
-      <RecipeCard recipe={longDescriptionRecipe} imageSizes="100vw" />,
-    );
+    render(<RecipeCard recipe={longDescriptionRecipe} imageSizes="100vw" />);
 
     const truncatedDescription = screen.getByText(/A{157}\.\.\./);
     expect(truncatedDescription).toBeInTheDocument();
   });
 
   it("renders the correct styles for hover state", () => {
-    const { container } = customRender(
+    const { container } = render(
       <RecipeCard recipe={mockRecipe} imageSizes="100vw" />,
     );
